Fix for 64-bit platforms

+++ RAW/AliRawReaderDate.cxx
@@ -85,7 +85,7 @@ AliRawReaderDate::AliRawReaderDate(
   while (fread(&header, 1, headerSize, fFile) == headerSize) {
     if (eventNumber == 0) {
       UChar_t* buffer = new UChar_t[header.eventSize];
-      fseek(fFile, -headerSize, SEEK_CUR);
+      fseek(fFile, -(long)headerSize, SEEK_CUR);
       if (fread(buffer, 1, header.eventSize, fFile) != header.eventSize) break;
       fEvent = (eventHeaderStruct*) buffer;
       break;
@@ -527,7 +527,7 @@ Bool_t AliRawReaderDate::NextEvent()
       continue;
     }
     UChar_t* buffer = new UChar_t[header.eventSize];
-    fseek(fFile, -headerSize, SEEK_CUR);
+    fseek(fFile, -(long)headerSize, SEEK_CUR);
     if (fread(buffer, 1, header.eventSize, fFile) != header.eventSize) {
       Error("NextEvent", "could not read event from file");
       delete[] buffer;
